fix: returns 404 instead of 401 missing headers
This occured when we didn't sent bearer token and tried to access an invalid endpoint.
The reason is because the grant middleware was returning error by not being able to extract the grants, not it's only
returning an empty set of grants. This causes that protected endpoints with fail dua a especific grant missing even
when the token is not present.